requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.14.0
transformers==4.35.2
openai==1.3.5
python-dateutil==2.8.2
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.6
